refactor: group domain modules under src/lib/{quran,analytics,habits}
