+++ src/lib/analytics/server-analytics.ts
@@ -1,0 +1,49 @@
+/**
+ * Nawaetu - Islamic Habit Tracker
+ * Copyright (C) 2026 Hadian Rahmat
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+export interface WebVitalMetric {
+    id: string;
+    name: string;
+    startTime: number;
+    value: number;
+    label: 'web-vital' | 'custom';
+}
+
+export interface AnalyticsProvider {
+    trackMetric(metric: WebVitalMetric): Promise<void>;
+}
+
+class ConsoleAnalyticsProvider implements AnalyticsProvider {
+    async trackMetric(metric: WebVitalMetric): Promise<void> {
+        console.info(`[Analytics] Web Vital: ${metric.name}`, metric);
+    }
+}
+
+// Factory to get the configured provider
+export function getAnalyticsProvider(): AnalyticsProvider {
+    // In future, return different provider based on config
+    return new ConsoleAnalyticsProvider();
+}
+
+/**
+ * Tracks a web vital metric using the configured analytics provider.
+ */
+export async function trackMetric(metric: WebVitalMetric): Promise<void> {
+    const provider = getAnalyticsProvider();
+    await provider.trackMetric(metric);
+}
